Flytt hentPersonidenterGjeldendeForBegrunnelse til BegrunnelseUtil og legg til beskrivelse på hentPersonerForUtvidetOgSmåbarnstillegg
